+++ crates/chia-sdk-driver/src/layers.rs
@@ -1,9 +1,9 @@
 mod nft_ownership_layer;
 mod nft_state_layer;
 mod singleton_layer;
-mod standard_layer;
+mod transparent_layer;
 
 pub use nft_ownership_layer::*;
 pub use nft_state_layer::*;
 pub use singleton_layer::*;
-pub use standard_layer::*;
+pub use transparent_layer::*;
