+++ crates/chia-sdk-driver/src/puzzles.rs
@@ -1,9 +1,7 @@
 mod cat;
 mod did;
-mod nft;
 mod singleton;
 
 pub use cat::*;
 pub use did::*;
-pub use nft::*;
 pub use singleton::*;
