--- crates/chia-sdk-driver/src/puzzles/nft.rs
@@ -1,5 +1,0 @@
-mod mint_nft;
-mod nft_spend;
-
-pub use mint_nft::*;
-pub use nft_spend::*;
